requirements.txt
deepface~=0.0.75
fastapi~=0.78.0
python-multipart
Pillow~=9.2.0
numpy~=1.23.0